+++ src/headless/plugins/omemo/device.js
@@ -3,12 +3,15 @@ import log from '@converse/log';
 import _converse from '../../shared/_converse.js';
 import api from '../../shared/api/index.js';
 import converse from '../../shared/api/public.js';
-import { IQError } from "../../shared/errors.js";
-import { UNDECIDED } from "./constants.js";
-import { parseBundle } from "./parsers.js";
+import { IQError } from '../../shared/errors.js';
+import { UNDECIDED } from './constants.js';
+import { parseBundle } from './parsers.js';
 
 const { Strophe, sizzle, stx, u } = converse.env;
 
+/**
+ * @extends {Model<import('./types').DeviceAttributes>}
+ */
 class Device extends Model {
     defaults() {
         return {
@@ -18,11 +21,11 @@ class Device extends Model {
     }
 
     /**
-     * @returns {import('./types').PreKey}
+     * @returns {import('./types').CounterpartyPreKey}
      */
     getRandomPreKey() {
         // XXX: assumes that the bundle has already been fetched
-        const bundle = this.get("bundle");
+        const bundle = this.get('bundle');
         return bundle.prekeys[u.getRandomInt(bundle.prekeys.length)];
     }
 
@@ -34,11 +37,11 @@ class Device extends Model {
      * @returns {Promise<import('./types').Bundle>}
      */
     async fetchBundleFromServer() {
-        const bare_jid = _converse.session.get("bare_jid");
+        const bare_jid = _converse.session.get('bare_jid');
         const stanza = stx`
-            <iq type="get" from="${bare_jid}" to="${this.get("jid")}" xmlns="jabber:client">
+            <iq type="get" from="${bare_jid}" to="${this.get('jid')}" xmlns="jabber:client">
                 <pubsub xmlns="${Strophe.NS.PUBSUB}">
-                    <items node="${Strophe.NS.OMEMO_BUNDLES}:${this.get("id")}"/>
+                    <items node="${Strophe.NS.OMEMO_BUNDLES}:${this.get('id')}"/>
                 </pubsub>
             </iq>`;
 
@@ -46,20 +49,20 @@ class Device extends Model {
         try {
             iq = await api.sendIQ(stanza);
         } catch (iq) {
-            log.error(`Could not fetch bundle for device ${this.get("id")} from ${this.get("jid")}`);
+            log.error(`Could not fetch bundle for device ${this.get('id')} from ${this.get('jid')}`);
             log.error(iq);
             if (iq && iq.querySelector('error')) {
                 throw new IQError('Could not fetch bundle', iq);
             }
             return null;
         }
-        if (iq.querySelector("error")) {
-            throw new IQError("Could not fetch bundle", iq);
+        if (iq.querySelector('error')) {
+            throw new IQError('Could not fetch bundle', iq);
         }
-        const publish_el = sizzle(`items[node="${Strophe.NS.OMEMO_BUNDLES}:${this.get("id")}"]`, iq).pop();
+        const publish_el = sizzle(`items[node="${Strophe.NS.OMEMO_BUNDLES}:${this.get('id')}"]`, iq).pop();
         const bundle_el = sizzle(`bundle[xmlns="${Strophe.NS.OMEMO}"]`, publish_el).pop();
         const bundle = parseBundle(bundle_el);
-        this.save("bundle", bundle);
+        this.save('bundle', bundle);
         return bundle;
     }
 
@@ -69,8 +72,8 @@ class Device extends Model {
      * @returns {Promise<import('./types').Bundle>}
      */
     getBundle() {
-        if (this.get("bundle")) {
-            return Promise.resolve(this.get("bundle"));
+        if (this.get('bundle')) {
+            return Promise.resolve(this.get('bundle'));
         } else {
             return this.fetchBundleFromServer();
         }