+++ src/headless/plugins/omemo/store.js
@@ -3,14 +3,15 @@ import log from '@converse/log';
 import _converse from '../../shared/_converse.js';
 import converse from '../../shared/api/public.js';
 import api from '../../shared/api/index.js';
+import { getCrypto } from './crypto.js';
 import { getDeviceList } from './utils.js';
 
 const { Strophe, stx, u } = converse.env;
 
+/**
+ * @extends {Model<import('./types').OMEMOStoreAttributes>}
+ */
 class OMEMOStore extends Model {
-    /**
-     * @typedef {Window & globalThis & {libsignal: any} } WindowWithLibsignal
-     */
     get Direction() {
         return {
             SENDING: 1,
@@ -19,67 +20,71 @@ class OMEMOStore extends Model {
     }
 
     /**
-     * @returns {Promise<import('./types').KeyPair>}
+     * @returns {import('libomemo.js').KeyPair}
      */
     getIdentityKeyPair() {
         const keypair = this.get('identity_keypair');
-        return Promise.resolve({
-            'privKey': u.base64ToArrayBuffer(keypair.privKey),
-            'pubKey': u.base64ToArrayBuffer(keypair.pubKey),
-        });
-    }
-
-    getLocalRegistrationId() {
-        return Promise.resolve(parseInt(this.get('device_id'), 10));
+        return {
+            privKey: u.base64ToArrayBuffer(keypair.privKey),
+            pubKey: u.base64ToArrayBuffer(keypair.pubKey),
+        };
     }
 
     /**
-     * @param {string} identifier
+     * @returns {number}
+     */
+    getLocalRegistrationId() {
+        return parseInt(this.get('device_id'), 10);
+    }
+
+    /**
+     * @param {string} address
      * @param {ArrayBuffer} identity_key
      * @param {unknown} _direction
+     * @returns {boolean}
      */
-    isTrustedIdentity(identifier, identity_key, _direction) {
-        if (identifier === null || identifier === undefined) {
+    isTrustedIdentity(address, identity_key, _direction) {
+        if (address === null || address === undefined) {
             throw new Error("Can't check identity key for invalid key");
         }
         if (!(identity_key instanceof ArrayBuffer)) {
             throw new Error('Expected identity_key to be an ArrayBuffer');
         }
-        const trusted = this.get('identity_key' + identifier);
+        const trusted = this.get('identity_key' + address);
         if (trusted === undefined) {
-            return Promise.resolve(true);
+            return true;
         }
-        return Promise.resolve(u.arrayBufferToBase64(identity_key) === trusted);
+        return u.arrayBufferToBase64(identity_key) === trusted;
     }
 
     /**
-     * @param {string} identifier
+     * @param {string} address
+     * @returns {ArrayBuffer}
      */
-    loadIdentityKey(identifier) {
-        if (identifier === null || identifier === undefined) {
-            throw new Error("Can't load identity_key for invalid identifier");
+    loadIdentityKey(address) {
+        if (address === null || address === undefined) {
+            throw new Error("Can't load identity_key for invalid address");
         }
-        return Promise.resolve(u.base64ToArrayBuffer(this.get('identity_key' + identifier)));
+        return u.base64ToArrayBuffer(this.get('identity_key' + address));
     }
 
     /**
-     * @param {string} identifier
+     * @param {string} address
      * @param {string} identity_key
+     * @returns {boolean}
      */
-    saveIdentity(identifier, identity_key) {
-        if (identifier === null || identifier === undefined) {
-            throw new Error("Can't save identity_key for invalid identifier");
+    saveIdentity(address, identity_key) {
+        if (address === null || address === undefined) {
+            throw new Error("Can't save identity_key for invalid address");
         }
-        const { libsignal } = /** @type WindowWithLibsignal */ (window);
-        const address = new libsignal.SignalProtocolAddress.fromString(identifier);
-        const existing = this.get('identity_key' + address.getName());
+        const existing = this.get('identity_key' + address);
         const b64_idkey = u.arrayBufferToBase64(identity_key);
-        this.save('identity_key' + address.getName(), b64_idkey);
+        this.save('identity_key' + address, b64_idkey);
 
         if (existing && b64_idkey !== existing) {
-            return Promise.resolve(true);
+            return true;
         } else {
-            return Promise.resolve(false);
+            return false;
         }
     }
 
@@ -102,17 +107,17 @@ class OMEMOStore extends Model {
     }
 
     /**
-     * @param {string} key_id
-     * @param {import('./types').KeyPair} key_pair
+     * @param {number} key_id
+     * @param {import('libomemo.js').KeyPair} key_pair
      */
     storePreKey(key_id, key_pair) {
-        const prekey = {};
-        prekey[key_id] = {
-            'pubKey': u.arrayBufferToBase64(key_pair.pubKey),
-            'privKey': u.arrayBufferToBase64(key_pair.privKey),
+        const prekey = {
+            [key_id]: {
+                privKey: u.arrayBufferToBase64(key_pair.privKey),
+                pubKey: u.arrayBufferToBase64(key_pair.pubKey),
+            },
         };
         this.save('prekeys', Object.assign(this.getPreKeys(), prekey));
-        return Promise.resolve();
     }
 
     /**
@@ -127,83 +132,71 @@ class OMEMOStore extends Model {
 
     /**
      * @param {string} _key_id
-     * @returns {Promise<import('./types').KeyPair|void>}
+     * @returns {import('libomemo.js').KeyPair|void}
      */
     loadSignedPreKey(_key_id) {
         const res = this.get('signed_prekey');
         if (res) {
-            return Promise.resolve({
-                'privKey': u.base64ToArrayBuffer(res.privKey),
-                'pubKey': u.base64ToArrayBuffer(res.pubKey),
-            });
+            return {
+                privKey: u.base64ToArrayBuffer(res.privKey),
+                pubKey: u.base64ToArrayBuffer(res.pubKey),
+            };
         }
-        return Promise.resolve();
     }
 
     /**
-     * @param {import('./types').SignedPreKey} spk
+     * @param {import('libomemo.js').SignedPreKey} spk
      */
     storeSignedPreKey(spk) {
         if (typeof spk !== 'object') {
-            // XXX: We've changed the signature of this method from the
-            // example given in InMemorySignalProtocolStore.
-            // Should be fine because the libsignal code doesn't
-            // actually call this method.
             throw new Error('storeSignedPreKey: expected an object');
         }
         this.save('signed_prekey', {
-            'id': spk.keyId,
-            'privKey': u.arrayBufferToBase64(spk.keyPair.privKey),
-            'pubKey': u.arrayBufferToBase64(spk.keyPair.pubKey),
-            // XXX: The InMemorySignalProtocolStore does not pass
-            // in or store the signature, but we need it when we
-            // publish our bundle and this method isn't called from
-            // within libsignal code, so we modify it to also store
-            // the signature.
-            'signature': u.arrayBufferToBase64(spk.signature),
+            id: spk.keyId,
+            privKey: u.arrayBufferToBase64(spk.keyPair.privKey),
+            pubKey: u.arrayBufferToBase64(spk.keyPair.pubKey),
+            signature: u.arrayBufferToBase64(spk.signature),
         });
-        return Promise.resolve();
     }
 
     /**
-     * @param {string} key_id
+     * @param {number} key_id
      */
     removeSignedPreKey(key_id) {
         if (this.get('signed_prekey')['id'] === key_id) {
             this.unset('signed_prekey');
             this.save();
         }
-        return Promise.resolve();
     }
 
     /**
-     * @param {string} identifier
+     * @param {string} address
      */
-    loadSession(identifier) {
-        return Promise.resolve(this.get('session' + identifier));
+    loadSession(address) {
+        return Promise.resolve(this.get('session' + address));
     }
 
     /**
-     * @param {string} identifier
+     * @param {string} address
      * @param {object} record
      */
-    storeSession(identifier, record) {
-        return Promise.resolve(this.save('session' + identifier, record));
+    storeSession(address, record) {
+        return Promise.resolve(this.save('session' + address, record));
     }
 
     /**
-     * @param {string} identifier
+     * @param {string} address
      */
-    removeSession(identifier) {
-        return Promise.resolve(this.unset('session' + identifier));
+    removeSession(address) {
+        return Promise.resolve(this.unset('session' + address));
     }
 
     /**
-     * @param {string} [identifier='']
+     * @param {string} [address='']
      */
-    removeAllSessions(identifier='') {
+    removeAllSessions(address = '') {
         const keys = Object.keys(this.attributes).filter((key) =>
-            key.startsWith('session' + identifier) ? key : false
+            key.startsWith('session' + address) ? key : false,
         );
         const attrs = {};
         keys.forEach((key) => {
@@ -219,11 +212,12 @@ class OMEMOStore extends Model {
         const item = stx`
             <item>
                 <bundle xmlns="${Strophe.NS.OMEMO}">
-                    <signedPreKeyPublic signedPreKeyId="${signed_prekey.id}">${signed_prekey.pubKey}</signedPreKeyPublic>
+                    <signedPreKeyPublic signedPreKeyId="${signed_prekey.id}"
+                        >${signed_prekey.pubKey}</signedPreKeyPublic>
                     <signedPreKeySignature>${signed_prekey.signature}</signedPreKeySignature>
                     <identityKey>${this.get('identity_keypair').pubKey}</identityKey>
                     <prekeys>${Object.values(this.get('prekeys')).map(
-                        (prekey, id) => stx`<preKeyPublic preKeyId="${id}">${prekey.pubKey}</preKeyPublic>`
+                        (prekey, id) => stx`<preKeyPublic preKeyId="${id}">${prekey.pubKey}</preKeyPublic>`,
                     )}
                     </prekeys>
                 </bundle>
@@ -233,12 +227,11 @@ class OMEMOStore extends Model {
     }
 
     async generateMissingPreKeys() {
-        const { libsignal } = /** @type WindowWithLibsignal */ (window);
-        const { KeyHelper } = libsignal;
+        const { KeyHelper } = await getCrypto();
 
         const prekeyIds = Object.keys(this.getPreKeys());
         const missing_keys = Array.from({ length: _converse.NUM_PREKEYS }, (_, id) => id.toString()).filter(
-            (id) => !prekeyIds.includes(id)
+            (id) => !prekeyIds.includes(id),
         );
 
         if (missing_keys.length < 1) {
@@ -274,8 +267,7 @@ class OMEMOStore extends Model {
      */
     async generatePreKeys() {
         const amount = _converse.NUM_PREKEYS;
-        const { libsignal } = /** @type WindowWithLibsignal */ (window);
-        const { KeyHelper } = libsignal;
+        const { KeyHelper } = await getCrypto();
         const keys = await Promise.all([...Array(amount).keys()].map((id) => KeyHelper.generatePreKey(id)));
 
         keys.forEach((k) => this.storePreKey(k.keyId, k.keyPair));
@@ -295,14 +287,14 @@ class OMEMOStore extends Model {
      * even if we're offline at that time.
      */
     async generateBundle() {
-        const { libsignal } = /** @type WindowWithLibsignal */ (window);
+        const { KeyHelper } = await getCrypto();
 
         // The first thing that needs to happen if a client wants to
         // start using OMEMO is they need to generate an IdentityKey
         // and a Device ID.
 
         // The IdentityKey is a Curve25519 public/private Key pair.
-        const identity_keypair = await libsignal.KeyHelper.generateIdentityKeyPair();
+        const identity_keypair = await KeyHelper.generateIdentityKeyPair();
         const identity_key = u.arrayBufferToBase64(identity_keypair.pubKey);
 
         // The Device ID is a randomly generated integer between 1 and 2^31 - 1.
@@ -317,7 +309,7 @@ class OMEMOStore extends Model {
             },
         });
 
-        const signed_prekey = await libsignal.KeyHelper.generateSignedPreKey(identity_keypair, 0);
+        const signed_prekey = await KeyHelper.generateSignedPreKey(identity_keypair, 0);
         this.storeSignedPreKey(signed_prekey);
 
         const prekeys = await this.generatePreKeys();
@@ -351,7 +343,7 @@ class OMEMOStore extends Model {
                      * @param {unknown} resp
                      */
                     error: (_model, resp) => {
-                        log.warn(`Could restore OMEMO session, we'll generate a new one: ${resp}`);
+                        log.warn(`Could restoring OMEMO session, we'll generate a new one: ${resp}`);
                         this.generateBundle().then(resolve).catch(reject);
                     },
                 });
@@ -362,22 +354,19 @@ class OMEMOStore extends Model {
 }
 
 export async function generateDeviceID() {
-    /**
-     * @typedef {module:plugins-omemo-index.WindowWithLibsignal} WindowWithLibsignal
-     */
-    const { libsignal } = /** @type WindowWithLibsignal */ (window);
+    const { KeyHelper } = await getCrypto();
 
     /* Generates a device ID, making sure that it's unique */
     const bare_jid = _converse.session.get('bare_jid');
     const devicelist = await getDeviceList(bare_jid, true);
     const existing_ids = devicelist.devices.pluck('id');
-    let device_id = libsignal.KeyHelper.generateRegistrationId();
+    let device_id = KeyHelper.generateRegistrationId();
 
     // Before publishing a freshly generated device id for the first time,
     // a device MUST check whether that device id already exists, and if so, generate a new one.
     let i = 0;
     while (existing_ids.includes(device_id)) {
-        device_id = libsignal.KeyHelper.generateRegistrationId();
+        device_id = KeyHelper.generateRegistrationId();
         i++;
         if (i === 10) {
             throw new Error('Unable to generate a unique device ID');