+++ src/headless/plugins/omemo/types.ts
@@ -1,7 +1,13 @@
 import { MUCMessageAttributes } from '../../plugins/muc/types';
-import { MessageAttributes } from '../../shared/types';
+import { MessageAttributes, ModelAttributes } from '../../shared/types';
 
-export type PreKey = {
+declare global {
+    interface Window {
+        libomemo?: Record<string, unknown>;
+    }
+}
+
+export type CounterpartyPreKey = {
     id: number;
     key: string;
 };
@@ -13,18 +19,33 @@ export type Bundle = {
         public_key: string;
         signature: string;
     };
-    prekeys: PreKey[];
+    prekeys: CounterpartyPreKey[];
+    fingerprint?: string;
 };
 
-export type KeyPair = {
+export type DeviceAttributes = ModelAttributes & {
+    id: number;
+    jid: string;
+    bundle?: Bundle;
+    trusted: 0 | 1 | -1;
+    active: boolean;
+};
+
+type SerializedKeyPair = {
     pubKey: string;
     privKey: string;
 };
 
-export type SignedPreKey = {
-    keyId: string;
-    keyPair: KeyPair;
-    signature: ArrayBuffer;
+type SerializedPreKey = {
+    id: number;
+    privKey: string;
+    pubKey: string;
+    signature: string;
+};
+
+export type OMEMOStoreAttributes = ModelAttributes & {
+    identity_keypair: SerializedKeyPair;
+    signed_prekey: SerializedPreKey;
 };
 
 export type EncryptedMessage = {
@@ -44,5 +65,3 @@ export type EncryptedMessageAttributes = {
 
 export type MUCMessageAttrsWithEncryption = MUCMessageAttributes & EncryptedMessageAttributes;
 export type MessageAttrsWithEncryption = MessageAttributes & EncryptedMessageAttributes;
-
-export type WindowWithLibsignal = Window & typeof globalThis & { libsignal: any };