+++ src/headless/types/plugins/omemo/constants.d.ts
@@ -1,6 +1,6 @@
-export const UNDECIDED: 0;
-export const TRUSTED: 1;
-export const UNTRUSTED: -1;
+/** @type {0} */ export const UNDECIDED: 0;
+/** @type {1} */ export const TRUSTED: 1;
+/** @type {-1} */ export const UNTRUSTED: -1;
 export const TAG_LENGTH: 128;
 export namespace KEY_ALGO {
     let name: string;